backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.12
email-validator==2.1.0
openai==1.12.0
httpx<0.25.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
feedparser==6.0.10
python-dateutil==2.8.2
lxml==4.9.3
